Sessions: use PHP's default session mechanism
